Update with PEP8 and rename parcer.py to parser.py
